мелкие правки

+++ migrations/20211117210222-project-session-table.js
@@ -17,7 +17,7 @@ module.exports = {
           }
         },
         sessionId: {
-          type: Sequelize.SMALLINT,
+          type: Sequelize.INTEGER,
           references: {
             model: 'Session',
             key: 'id'
